Add front squat calculator feat
Add front squat feat to assess the goal input

Feature compares goal relative to body weight to validate if the goal needs to increase.

Feature makes appropriate goal weight simple for a user. If the goal weight is significantly too low the feature breaks the program and the user must ajust the inputs.